[COLAB-2918] Create separate filter configuration
This removes dependencies on the filtesr from the WebConfig class, which created a circular dependency with feign
